requirements.txt
discord.py
python-dotenv
colorlog
rich
pynacl